server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
openai
pypdf